+++ sites/docs/src/lib/components/docs/pm-execute.svelte
@@ -1,0 +1,6 @@
+<script lang="ts">
+	import PMBlock from "./pm-block.svelte";
+	export let command: string;
+</script>
+
+<PMBlock type="execute" {command} />