+++ sites/docs/src/lib/registry/default/ui/card/card-header.svelte
@@ -9,6 +9,6 @@
 	}: PrimitiveDivAttributes = $props();
 </script>
 
-<div bind:this={ref} class={cn("flex flex-col space-y-1.5 p-6", className)} {...restProps}>
+<div bind:this={ref} class={cn("flex flex-col space-y-1.5 p-6 pb-0", className)} {...restProps}>
 	{@render children?.()}
 </div>