+++ sites/docs/src/lib/registry/default/ui/separator/separator.svelte
@@ -14,7 +14,7 @@
 	bind:ref
 	class={cn(
 		"bg-border shrink-0",
-		orientation === "horizontal" ? "h-[1px] w-full" : "h-full w-[1px]",
+		orientation === "horizontal" ? "h-[1px] w-full" : "min-h-full w-[1px]",
 		className
 	)}
 	{orientation}