+++ sites/docs/src/lib/registry/new-york/ui/card/card-content.svelte
@@ -9,6 +9,6 @@
 	}: PrimitiveDivAttributes = $props();
 </script>
 
-<div bind:this={ref} class={cn("p-6 pt-0", className)} {...restProps}>
+<div bind:this={ref} class={cn("p-6", className)} {...restProps}>
 	{@render children?.()}
 </div>