+++ sites/docs/src/routes/(app)/examples/forms/account/account-form.svelte
@@ -1,4 +1,4 @@
-<script lang="ts" module>
+<script lang="ts" context="module">
 	import { z } from "zod";
 
 	const languages = [
@@ -26,11 +26,11 @@
 		language: z.enum(languages.map((lang) => lang.value) as [Language, ...Language[]]),
 		dob: z
 			.string()
-			.datetime()
+			.datetime({ message: "" })
 			// we're setting it optional so the user can clear the date and we don't run into
 			// type issues, but we refine it to make sure it's not undefined
 			.optional()
-			.refine((date) => date !== undefined, "Please select a valid date."),
+			.refine((date) => date !== "", "Please select a valid date."),
 	});
 
 	export type AccountFormSchema = typeof accountFormSchema;